Configure: rethink overrides coming from the environment
We have a number of lines that add config attributes for the target
from environment variables, or defaults if the environment variable
doesnt exist.  These were made to depend on the original config
attribute when they really shouldn't, so let's reorganise things so
they get added to the config target directly before trying to reolve
inherited values.  This avoid unecessary computation, and possibly
complexity.