Configure: allow late processing of the resolved config target
Previously, we did the final processing of CODErefs in the selected
config target at the same time as resolving the chain of inheritance.
It's possible, though, that the processing of some attributes requires
configuration data that is figured out later, and that we therefore
need to delay running them until they're actually needed.

This is solved by tying %target to a special hash processing class
that only does the final attribute resolution when the value of that
particular attribute is asked for (which is when this attribute is
used as an rvalue, such as '$whatever = $target{foo}').

An example where this could come in handy is if some attribute ('foo',
say...) is computed based on the C compiler flavor.  We know that just
grepping for 'gcc' in some help output may not be enough...  and sure
enough, we do have a %predefined hash table that helps finding out
with better precision, but that is computed after $target{cc} is
resolved, so we end up being in a catch 22 with the previous scheme.
Now, the computation of that attribute ('foo') has a better chance of
being delayed until after $target{cc} has received its final value.