Separate EVP PKEY and the algorithm implementations
This means that the EVP layer cannot make *any* assumption on the
actual key contents and must therefore treat the implementation data
as unknown (i.e. a void *, which could also be an undefined structure)

This also means that the i2d and d2i functions for public keys must
move to the EVP_PKEY_ASN1_METHOD structure.