Rework building: add install functionality
Installation of most things is fairly straight forward, but involves a
lot of copied commands.  However, when it comes to installing
libraries, things become trickier, as different platforms have
different views on how to handle them, especially the shared
libraries.

We therefore want a set of simple functions that can handle the more
complicated situations in a platform specific manner, and help keep
the build file templates cleaner.